Refactor patent-browser.vue into smaller components for better organization and maintainability
